+++ custom-addons/sh_activities_management/doc/changelog.rst
@@ -32,3 +32,69 @@ Initial Release
 17.0.8(Date: 14th March 2024)
 =====================================
 [ADD] Dashboard feature added.
+
+17.0.9 (Date: 13th April 2024)
+---------------------------------------
+[ADD] add reminder field in the activity view.
+
+17.0.10 (Date: 16th April 2024)
+---------------------------------------
+[Fix] fix domain filter in activity type in dashboard.
+[Fix] Resolve the issue with the Activity Document Model Configuration to accurately showcase activities model by model, ensuring that only selected models are displayed in the dashboard.
+
+17.0.11 (Date: 30th April 2024)
+---------------------------------------
+[Fix] Add new python file to inherite mail.activity.schedule model and display fields in 'Schedule Activity' wizard and add _action_schedule_activities method to pass data added in schedule activity wizard.
+
+17.0.12 (Date: 9th May 2024)
+--------------------------------------
+[Add] Add Mark as done Button in Create Multiple activities feature.
+
+17.0.13 (Date: 24th July 2024)
+--------------------------------------
+[Fix] Add the 'sh_date_deadline' field in mail.activity.schedule and display the sh_date_deadline field in the Schedule Activity Form View and Activities Form View. 
+Fix the issue to display the "Activity Reminder" Popup notification.
+
+17.0.14 (Date: 1st August 2024)
+--------------------------------------
+[Fix] Fix the issue of display feedback using standard odoo chatter. ( in mail.activity model action_feedback())
+
+17.0.15 (Date: 14th August 2024)
+--------------------------------------
+[Fix] Fix the issue of display feedback using standard odoo chatter when Click 'Done & Schedule Next' Button. ( in mail.activity model _action_done())
+
+17.0.16 (Date: 6th November 2024)
+---------------------------------------
+[FIX] warning of required field
+
+17.0.17 (Date: 18th November 2024)
+-------------------------------------
+[FIX] error on mass done activities.
+
+17.0.18 (Date: 21st November 2024)
+-------------------------------------
+[FIX] issue of wrong counts.
+
+17.0.19 (Date: 12th December 2024)
+-------------------------------------
+[FIX] Access right issue fixed.
+
+17.0.20 (Date: 19th December 2024)
+-------------------------------------
+[FIX] Updated the code to filter data by due date and completion date when applying filters.
+
+17.0.21 (Date: 27th December 2024)
+-------------------------------------
+[FIX] warnings.
+
+17.0.22 (Date: 24th January 2025)
+-------------------------------------
+[FIX] Fixed the warning of use 'name_get' method deprected in mail activity _compute_res_name().
+
+17.0.23 (Date: 28th April 2025)
+-------------------------------------
+[FIX] activity delete issue fixed while recruitment application moved to the refuse.
+
+17.0.24 (Date: 6th May 2025)
+-------------------------------------
+[FIX] added fields to optional in tree view